Factor out test base
